refactor: unify prompt handling logic across translation services
