+++ src/modules/services/gemini.ts
@@ -1,6 +1,6 @@
-import { TranslateTask } from "../../utils/task";
-import { getPref } from "../../utils/prefs";
+import { getPref, transformPromptWithContext } from "../../utils";
 import { TranslateService } from "./base";
+import type { TranslateTask } from "../../utils/task";
 
 const translate = <TranslateService["translate"]>async function (data) {
   const apiURL = getPref("gemini.endPoint") as string;
@@ -10,40 +10,7 @@ const translate = <TranslateService["translate"]>async function (data) {
     langTo: string,
     sourceText: string,
   ) {
-    let prompt = getPref("gemini.prompt") as string;
-
-    // Add paper context if enabled and itemId is available
-    if (getPref("attachPaperContext") && data.itemId) {
-      const item = Zotero.Items.get(data.itemId);
-      const topItem = item ? Zotero.Items.getTopLevel([item])[0] : null;
-      if (topItem) {
-        let contextInfo = "";
-        const title = topItem.getField("title") as string;
-        const abstract = topItem.getField("abstractNote") as string;
-
-        if (title) {
-          contextInfo += `Paper Title: ${title}`;
-        }
-        if (abstract) {
-          contextInfo += title
-            ? `\n\nPaper Abstract: ${abstract}`
-            : `Paper Abstract: ${abstract}`;
-        }
-
-        if (contextInfo) {
-          // Insert context before the source text
-          prompt = prompt.replace(
-            "${sourceText}",
-            `Context from the academic paper:\n${contextInfo}\n\nText to translate: ${sourceText}`,
-          );
-        }
-      }
-    }
-
-    return prompt
-      .replaceAll("${langFrom}", langFrom)
-      .replaceAll("${langTo}", langTo)
-      .replaceAll("${sourceText}", sourceText);
+    return transformPromptWithContext("gemini.prompt", langFrom, langTo, sourceText, data);
   }
 
   function getGenContentAPI(data: Required<TranslateTask>) {
